Updating this sample to target a profile.
To do this, and make it easy to just run this sample, I had to update
Glide to target the `dev` branch. This should be a temporary state, and
once profiles become mainstream, we should update glide.yaml to target a
release tag again.

Tareting a profile here both helps how how to use it, and makes our
samples more resilient to breaking changes in the future.